requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Mail==0.9.1

# Database and Authentication
supabase==2.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
Flask-Login==0.6.3

# Task Queue and Caching
celery==5.3.6
redis==5.2.1
Flask-Caching==2.1.0

# API and Networking
requests==2.31.0
aiohttp==3.11.8
httpx==0.26.0
tweepy==4.14.0
python-twitter-v2==0.8.1

# Video Processing
opencv-python==4.9.0.80
moviepy==1.0.3
yt-dlp==2024.3.10
ffmpeg-python==0.2.0
Pillow==9.5.0
mediapipe==0.10.21
numpy==1.26.3
pydub>=0.25.1

# Web Framework Extensions
flask-cors==4.0.0
flask-session==0.5.0

# Web APIs
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.70.0

# Security and Monitoring
Flask-Talisman==1.1.0
Flask-SeaSurf==1.1.1
prometheus-flask-exporter==0.23.0
sentry-sdk[flask]==1.40.0

# Development and Testing
pytest==7.4.4
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Web Interface
Flask-WTF==1.2.1
Flask-Assets==2.1.0
Jinja2==3.1.3

# Performance and Optimization
uvicorn[standard]==0.30.6
fastapi==0.111.0
pydantic==2.9.2

# Utilities
python-slugify==8.0.4
pytz==2024.1
tqdm==4.66.3
python-magic==0.4.27

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Production
supervisor==4.2.5
newrelic==9.6.0

face_recognition
# Text generation
openai>=0.28.0,<1.0.0
tiktoken>=0.4.0,<0.6.0
git+https://github.com/rabilrbl/deepseek-api.git
google-generativeai>=0.3.0

# Utilities
validators>=0.22.0
pyyaml>=6.0.1
colorlog>=6.7.0
python-magic>=0.4.27